# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API HUB-UMKM Documentation",
        "description": "Documentation for HUB PHB UMKM",
        "contact": {
            "email": "jamal.apriadi@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Laravel Swagger API server"
        }
    ],
    "paths": {
...
        "/api/register-seller": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user (customer or seller)",
                "operationId": "e308b2b868b8a672ddeddb016b9d2359",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Andi"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "andi@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "seller",
                                            "customer"
                                        ],
                                        "example": "seller"
                                    },
                                    "shop_name": {
                                        "type": "string",
                                        "example": "Toko Andi"
                                    },
                                    "shop_url": {
                                        "type": "string",
                                        "example": "toko-andi"
                                    },
                                    "shop_description": {
                                        "type": "string",
                                        "example": "Menjual barang elektronik"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Seller registered successfully"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Andi"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "andi@example.com"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "seller"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "seller_info": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "shop_name": {
                                                    "type": "string",
                                                    "example": "Toko Andi"
                                                },
                                                "shop_url": {
                                                    "type": "string",
                                                    "example": "toko-andi"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Registration failed"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login for customer/seller and get access token",
                "operationId": "562f93a4a4d383c95ba1c9b6d7170190",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "andi@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 604800
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1Q..."
                                        },
                                        "refresh_token": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Password wrong / User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Password wrong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}